requirements.txt
fastapi
uvicorn[standard]
jinja2
python-dotenv
pydantic
dashscope
itsdangerous
cryptography
openai
